chore: remove Gemini 1.5 adapter
