+++ aidial_adapter_vertexai/chat/gemini/prompt/gemini_2.py
@@ -18,9 +18,6 @@ from aidial_adapter_vertexai.chat.gemini.processors import (
     get_video_processor,
 )
 from aidial_adapter_vertexai.chat.gemini.prompt.base import GeminiPromptGenAI
-from aidial_adapter_vertexai.chat.gemini.prompt.gemini_1_5 import (
-    get_usage_message,
-)
 from aidial_adapter_vertexai.chat.gemini.prompt.message import (
     GenAIMessageMerger,
 )
@@ -74,3 +71,19 @@ class Gemini_2_Prompt(GeminiPromptGenAI):
         return cls(
             conversation=conversation, tools=tools, static_tools=static_tools
         )
+
+
+def get_usage_message(exts: List[str]) -> str:
+    return f"""
+The application answers queries about attached documents.
+Attach documents and ask questions about them in the same message.
+
+Supported document extensions: {', '.join(exts)}.
+
+Examples of queries:
+- "Describe the picture" for one image,
+- "What is depicted in these images?", "Compare the images" for multiple images,
+- "Summarize the document" for a PDF,
+- "Transcribe the audio" for an audio file,
+- "What is happening in the video?" for a video.
+""".strip()
