+++ tests/unit_tests/test_endpoints.py
@@ -23,8 +23,6 @@ class TestCase:
 
 
 test_cases: List[TestCase] = [
-    TestCase(D.GEMINI_PRO_1_5_V2, True, True, False),
-    TestCase(D.GEMINI_FLASH_1_5_V2, True, True, False),
     TestCase(D.IMAGEN_005, True, True, True),
     TestCase(D.IMAGEN_3_GENERATE_001, True, True, True),
     TestCase(D.IMAGEN_3_GENERATE_002, True, True, True),
